+++ core/providers.go
@@ -14,7 +14,7 @@ func ProviderList() (providers map[string][]string, err error) {
 
 	providers = make(map[string][]string)
 
-	for _, provider := range []string{"CoinGate", "Coingecko"} {
+	for _, provider := range []string{"CoinGate", "Coingecko", "LiveCoinWatch"} {
 		providers[provider] = append(providers[provider], "default")
 	}