+++ providers/livecoinwatch/pricequery.go
@@ -1,0 +1,8 @@
+package livecoinwatch
+
+// PriceQuery encapsulates an API request for an exchange rate.
+type PriceQuery struct {
+	Currency string `json:"currency"`
+	Code     string `json:"code"`
+	Meta     bool   `json:"meta"`
+}